requirements.txt
selenium==4.11.2
requests==2.31.0
webdriver-manager==3.7.1
python-dotenv==1.0.0
Faker==19.3.0
pytest==7.4.0
pytest-xdist==3.3.1
allure-pytest==2.13.2